requirements.txt
ultralytics
opencv-python
numpy
python-telegram-bot
PyYAML